+++ src/app/components/PokemonList.tsx
@@ -1,26 +1,65 @@
 "use client";
 import "./PokemonList.scss";
 
-import { getPokemonList } from "@/server/pokemon/get";
+import { getPokemonList, PokemonDetailType } from "@/server/pokemon/get";
 import Image from "next/image";
 import Link from "next/link";
-import { useEffect } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useServerAction } from "zsa-react";
 
 const PokemonList = () => {
-  const {
-    isPending: isPendingPokemonList,
-    execute: executeGetPokemonList,
-    data: pokemonList,
-  } = useServerAction(getPokemonList /*, { initialData: {} }*/);
+  const [pokemonList, setPokemonList] = useState<PokemonDetailType[]>([]);
+  const [offset, setOffset] = useState(0);
+  const observerRef = useRef<IntersectionObserver | null>(null);
 
-  console.log({ isPendingPokemonList, pokemonList });
+  const { isPending: isPendingPokemonList, execute: executeGetPokemonList } =
+    useServerAction(getPokemonList, { persistDataWhilePending: true });
+
+  const fetchPokemonList = useCallback(
+    async (offset: number) => {
+      if (isPendingPokemonList) return;
+
+      const [pokemonListResponse, err] = await executeGetPokemonList({
+        limit: 18,
+        offset,
+      });
+
+      if (err) {
+        console.error(err);
+        return;
+      }
+
+      setPokemonList((prev) => [...prev, ...pokemonListResponse]);
+      setOffset((prev) => prev + 18);
+    },
+    [executeGetPokemonList, isPendingPokemonList]
+  );
 
   useEffect(() => {
-    executeGetPokemonList({ limit: 100, offset: 600 });
-  }, [executeGetPokemonList]);
+    fetchPokemonList(offset);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
-  if (isPendingPokemonList || !pokemonList) {
+  useEffect(() => {
+    if (observerRef.current) observerRef.current.disconnect();
+
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting && !isPendingPokemonList) {
+          fetchPokemonList(offset);
+        }
+      },
+      { root: null, rootMargin: "0px", threshold: 0.5 }
+    );
+
+    const listener = document.querySelector("#listener");
+    if (listener) observer.observe(listener);
+    observerRef.current = observer;
+
+    return () => observer.disconnect();
+  }, [fetchPokemonList, isPendingPokemonList, offset]);
+
+  if (!pokemonList) {
     return <div>Loading...</div>;
   }
 
@@ -43,10 +82,11 @@ const PokemonList = () => {
           </div>
         </Link>
       ))}
+      <div id="listener" className="listener">
+        {isPendingPokemonList && "Loading more pokemon..."}
+      </div>
     </div>
   );
-
-  // return JSON.stringify(pokemonList);
 };
 
 export default PokemonList;