+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - NELO-5: Implement first approach of `<PokemonList />`
 - NELO-6: Configure Pokemon background color according to their type.
 - NELO-7: Implement `getPokemonById` Server Action and implement it in Pokemon Details page using `<PokemonDetails>` component
+- NELO-8: Implement IntersectionObserver in PokemonList to enable Infinite Scroll effect